// src/app/modules/hcm/shared/models/laborRelationship/family-burden.ts
export class FamilyBurden{

    idFamilyBurden: number =-1;
    idEmployee: number =-1;
    idLaborRelationship: number = -1;
    idLaborRelationshipxFamilyBurden: number = -1;
    idDocumentType: number = -1;
    idKinship: number = -1;
    documentNumber: string ="";
    firstName: string ="";
    lastName: string ="";
    observation: string ="";
    gender: string = "";
    workFlag: boolean = false;
    impairmentFlag: boolean = false;
    studyFlag: boolean = false;
    declaredFlag: boolean = false;
    active: boolean = false;
    createdByUserId: number = -1;
    updatedByUserId: number = -1;
    birthDate: string;
    registrationDate: string;
   
}